add Screen higher order component
