+++ screens/LinksScreen.js
@@ -1,17 +1,17 @@
 import React from 'react';
-import { ScrollView, StyleSheet } from 'react-native';
 import { ExpoLinksView } from '@expo/samples';
+import Screen from '../components/Screen';
 
 export default function LinksScreen() {
   return (
-    <ScrollView style={styles.container}>
+    <Screen>
       {/**
        * Go ahead and delete ExpoLinksView and replace it with your content;
        * we just wanted to provide you with some helpful links.
        */}
       <ExpoLinksView />
 
-    </ScrollView>
+    </Screen>
   );
 }
 
@@ -19,10 +19,3 @@ LinksScreen.navigationOptions = {
   title: 'Links!',
 };
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    paddingTop: 15,
-    backgroundColor: '#fff',
-  },
-});
