+++ screens/SermonArchiveScreen.js
@@ -2,7 +2,6 @@ import * as WebBrowser from 'expo-web-browser';
 import React, { useEffect } from 'react';
 import {
   Image,
-  Button,
   Platform,
   ScrollView,
   StyleSheet,
@@ -10,18 +9,16 @@ import {
   TouchableOpacity,
   View,
   ImageBackground,
-  WebView,
   FlatList,
   TouchableHighlight,
   ActivityIndicator,
 } from 'react-native';
-import Touchable from 'react-native-platform-touchable';
 import { Ionicons } from '@expo/vector-icons';
 import { connect } from 'react-redux';
 
 import { getSermons } from '../actions/sermonArchives';
-import { MonoText } from '../components/StyledText';
 import  VideoItem  from '../components/VideoItem';
+import Screen from '../components/Screen';
 
 
 function SermonArchiveScreen ({
@@ -43,72 +40,55 @@ function SermonArchiveScreen ({
   );
 
   return (
-    <View style={styles.container}>
-      <ImageBackground source={require('../assets/images/background-image.jpg')} style={{width: '100%', height: '100%', resizeMode: 'cover'}}>
-        <ScrollView
-          style={styles.container}
-          contentContainerStyle={styles.contentContainer}>
-          <View style={styles.welcomeContainer}>
-            <TouchableOpacity onPress={handleLogoPress}>
-              <Image
-                source={
-                  __DEV__
-                    ? require('../assets/images/kt-logo.png')
-                    : require('../assets/images/kt-logo.png')
-                }
-                style={styles.welcomeImage}
-              />
-            </TouchableOpacity>
-            <Text style={styles.pageTitleText}>Latest Sermons</Text>
-          </View>
-          { 
-            booksRT !== undefined 
-              ? (
-              <View style={{flex: 1, paddingTop: 5, alignItems: 'center'}}>
+    <Screen
+      title="Latest Sermons"
+    >
+      { 
+        booksRT !== undefined 
+          ? (
+          <View style={{flex: 1, paddingTop: 5, alignItems: 'center'}}>
 
-                <FlatList
-                  style={{width:'100%'}}
-                  data={booksRT}
-                  renderItem={({ item }) => {
-                    return (
-                      <VideoItem
-                        item = {item}
-                      />
-                    )
-                  }}
-                  keyExtractor={(item, index) => index.toString()}
-                />
-                <View style={styles.container, {alignItems: 'center', flexDirection:'row', marginBottom:30}}>
-                <TouchableHighlight
-                  style={styles.buttonSquare}
-                  onPress={handlePressSeeMoreVideos}
-                  underlayColor='rgba(250, 168, 127, 1)'>
-                      <Text style={styles.homeLinkText}>See more{"\n"}on our website</Text>
-                </TouchableHighlight>
-                <TouchableHighlight
-                    style={styles.buttonSquare}
-                    onPress={() => goBack()} title="Go back"
-                    underlayColor='rgba(250, 168, 127, 1)'>
-                      <Text style={styles.homeLinkText}>Go{"\n"}Back{"\n"}
-                        <Ionicons
-                          style={{display:'flex'}}
-                          size={32}
-                          name={Platform.OS === 'ios'
-                          ? 'ios-arrow-back' : 'md-arrow-back'}/>
-                      </Text>
-                  </TouchableHighlight>
-                </View>
-              </View>
-              ) 
-            : (
-              <View style={{ flex: 1, padding: 20 }}>
-                <ActivityIndicator size="large" color="#ffffff"/>
-              </View>
-            )
-          }
-    </ScrollView>
-    </ImageBackground>
-    </View>
+            <FlatList
+              style={{width:'100%'}}
+              data={booksRT}
+              renderItem={({ item }) => {
+                return (
+                  <VideoItem
+                    item = {item}
+                  />
+                )
+              }}
+              keyExtractor={(item, index) => index.toString()}
+            />
+            <View style={styles.container, {alignItems: 'center', flexDirection:'row', marginBottom:30}}>
+            <TouchableHighlight
+              style={styles.buttonSquare}
+              onPress={handlePressSeeMoreVideos}
+              underlayColor='rgba(250, 168, 127, 1)'>
+                  <Text style={styles.homeLinkText}>See more{"\n"}on our website</Text>
+            </TouchableHighlight>
+            <TouchableHighlight
+                style={styles.buttonSquare}
+                onPress={() => goBack()} title="Go back"
+                underlayColor='rgba(250, 168, 127, 1)'>
+                  <Text style={styles.homeLinkText}>Go{"\n"}Back{"\n"}
+                    <Ionicons
+                      style={{display:'flex'}}
+                      size={32}
+                      name={Platform.OS === 'ios'
+                      ? 'ios-arrow-back' : 'md-arrow-back'}/>
+                  </Text>
+              </TouchableHighlight>
+            </View>
+          </View>
+          ) 
+        : (
+          <View style={{ flex: 1, padding: 20 }}>
+            <ActivityIndicator size="large" color="#ffffff"/>
+          </View>
+        )
+      }
+    </Screen>
   );
 }
 
@@ -133,22 +113,6 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: 'rgba(137, 167, 165, 0.8)',
-  },
-  contentContainer: {
-    marginTop: 50,
-    flex: 1,
-  },
-  welcomeContainer: {
-    alignItems: 'center',
-    marginTop: 0,
-    marginBottom: 20,
-  },
-  welcomeImage: {
-    width: 100,
-    height: 80,
-    resizeMode: 'contain',
-    marginTop: 3,
-    marginLeft: 0,
   },
   buttonSquare: {
     padding: 15,
@@ -197,16 +161,6 @@ const styles = StyleSheet.create({
     textAlign: 'center',
     textTransform: 'uppercase',
     fontWeight: 'bold',
-  },
-  pageTitleText: {
-    color: '#fff',
-    fontSize: 12,
-    margin: 'auto',
-    paddingTop:20,
-    letterSpacing:3,
-    textAlign: 'center',
-    textTransform: 'uppercase',
-    fontWeight: 'normal',
   },
   tabBarInfoContainer: {
     position: 'absolute',
