+++ screens/SettingsScreen.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import { ExpoConfigView } from '@expo/samples';
+import Screen from '../components/Screen';
 
 export default function SettingsScreen() {
-  /**
-   * Go ahead and delete ExpoConfigView and replace it with your content;
-   * we just wanted to give you a quick view of your config.
-   */
-  return <ExpoConfigView />;
+  return (
+    <Screen>
+      <ExpoConfigView />
+    </Screen>
+  );
 }
 
 SettingsScreen.navigationOptions = {
